write metric test
