+++ controllers/k8sclient/interface.go
@@ -2,6 +2,7 @@
 package k8sclient
 
 import (
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
@@ -14,4 +15,6 @@ type Interface interface {
 	kubernetes.Interface
 	dynamic.Interface
 	Patch(gvr schema.GroupVersionResource, objNamespace string, objName string, by []byte) (*unstructured.Unstructured, error)
+	GetSecret(string, string) (*corev1.Secret, error)
+	UpdateSecret(*corev1.Secret) (*corev1.Secret, error)
 }
